% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{listIps}
\alias{listIps}
\title{Get a list of your private and public ip}
\usage{
listIps(version = 4L, throw = TRUE)
}
\arguments{
\item{version}{Integer. Ip version \code{4} or \code{6}.}

\item{throw}{Logical. In case your ip addresses could not be resolved
should an error be thrown?}
}
\value{
Named list; if \code{!throw} \code{NULL} on failure.
}
\description{
Get a list of your private and public ip
}

