% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{listV4}
\alias{listV4}
\title{Get a list of your private and public IPv4}
\usage{
listV4(throw = TRUE)
}
\arguments{
\item{throw}{Logical. In case your Ip addresses could not be resolved 
should an error be thrown?}
}
\value{
Named list; if \code{!throw} \code{NULL} on failure.
}
\description{
Get a list of your private and public IPv4
}

