% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{privateIp}
\alias{privateIp}
\title{Get your private ip address}
\usage{
privateIp(version = 4L, throw = TRUE)
}
\arguments{
\item{version}{Integer. Ip version \code{4} or \code{6}. DISABLED!}

\item{throw}{Logical. In case your ip address could not be resolved
should an error be thrown?}
}
\value{
Character. Private ip address; if \code{!throw} \code{NULL}
on failure.
}
\description{
Get your private ip address
}
\details{
Calls \code{ipconfig} on Windows, \code{ifconfig} on OSX, in both
cases followed by regex parsing; \code{hostname -I} on Linux with out any
further parsing.
}

