% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{privateV4}
\alias{privateV4}
\title{Get your private IP address as v4}
\usage{
privateV4(throw = TRUE)
}
\arguments{
\item{throw}{Logical. In case your Ip address could not be resolved 
should an error be thrown?}
}
\value{
Character. Private ip address; if \code{!throw} \code{NULL} 
on failure.
}
\description{
Get your private IP address as v4
}
\details{
Calls \code{ipconfig} on Windows and \code{ifconfig} on all
other OS.
}

