% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{publicIp}
\alias{publicIp}
\title{Get your public ip address}
\usage{
publicIp(version = 4L, throw = TRUE)
}
\arguments{
\item{version}{Integer. Ip version \code{4} or \code{6}. Caution: see below.}

\item{throw}{Logical. In case your ip address could not be resolved
should an error be thrown? DISABLED!}
}
\value{
Character. Public ip address; if \code{!throw} \code{NULL}
on failure.
}
\description{
Get your public ip address
}
\details{
CAUTION: forcing a specific version is unreliable - ipv4 and ipv6
seem to be mutually exclusively resolve... In case your desired version
could not be resolved, you can however get the version you did not demand.
Hehe. Queryies ichanhazip.com through SSL.
}

