% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{publicV4}
\alias{publicV4}
\title{Get your public IP address as v4}
\usage{
publicV4(throw = TRUE)
}
\arguments{
\item{throw}{Logical. In case your Ip address could not be resolved 
should an error be thrown?}
}
\value{
Character. Public ip address; if \code{!throw} \code{NULL} 
on failure.
}
\description{
Get your public IP address as v4
}
\details{
Queryies ichanhazip.com through SSL for your public Ip v4.
}

